Rename and move CBPVToCPSTypes to own file
