+++ src/test/scala/ilc/feature/let/Instantiations.scala
@@ -22,7 +22,7 @@ trait Instantiations {
   def buildBacchusWithLetSystem(doCSE_ : Boolean, copyPropagation_ : Boolean, partialApplicationsAreSpecial_ : Boolean) =
     new language.LetLanguage with let.ShowTerms with let.ANormalFormAdapter with inference.InferenceTestHelper
     //XXX added to also test CPS in worksheets
-    with let.CPS with cbpv.CBPVToCPS {
+    with let.CPS with cbpv.CBPVToCPSTypes {
       outer =>
       val aNormalizer = new ANormalFormStateful {
         val syntax: outer.type = outer
